Added controls so user can rotate and translate a planet
Added a controller class which checks for keyboard input at the moment
with future possibilities of adding mouse controls, etc. Acts as a proxy
pattern for user input. Also added jQuery for getting the input and
downloaded it because it was affecting load times when I was grabbing it
from google (probably due to my internet connection being slow but I was
testing frequently which meant redownloading frequently). Also modified
planet.js so it could cope with rotation and provide functions for
moving it (with the aim of making it more readable). Also refactored
some variable names by prefixing with a _ to indicate they are a class
member.