Implemented Tessellation
Implemented the CreateSphere method that tessellates the triangles into
4 smaller triangles. This takes in a patch of nine, aka it takes in nine
vertices at a time, and produces 9x4 vertices pushed into the array
called newVerts and returns this array on completion. This method can be
called multiple times in this manner:

for(tessellationDepth)
_verts = CreateSphere(_verts, radius);

this assumes there are some "base" vertices already in _verts which can
be tessellated. In this program the base vertices make up a diamond. It
tessellates the vertices in _verts then returns them and stores them in
_verts, then the loop goes around and it tessellates the previously
tessellated vertices. So to tessellate to a depth of 3 you run the for
loop 3 times and at the end the vertices in _verts will have been
tessellated 3 times. There is a maximum depth of tessellation you can do
but I believe this is machine dependent as the greater the depth the
greater the number of vertices and therefore the greater amount of
memory and processing power required.