Turn tessellated Diamond into Sphere
Added code to turn the tessallated diamond into a sphere. This is done
by calculating the length of each vector, normalising it and then
multiplying the normalised vector by the desired radius. The smoothness
of the sphere is determined by the tessellation depth which is set with
the variable _tessellationDepth. The higher this is the more slower load
times are though as more depth = more processing.