Added all planets and fixed positioning bug
Added all the planets of the solar system and fixed a bug that was
affecting the positioning of planets. The Y location was unchanged when
set in initialisation but worked when altered during runtime. This
turned out to be an issue related to the perspective. changed the max
from 100.0 to 1000.0 and everything works fine now.