Fixed orbits, put to scale (size/orbit), plus more
Fixed orbits so that the planets orbit around the sun approx. correctly.
Put the planets to scale and the speed of the orbits in rough
proportion. Also added controls to rotate the camera, zoom in and out
and speed up / slow down time. Also added shaders for every entity (sun,
planets etc.) currently they just shade them a colour, but later can be
used to add textures and bump mapping.